design: 컨테이너 컴포넌트 통합
