+++ src/pages/user/style.scss
@@ -1,14 +1,4 @@
-.container {
-    width: 100%;
-    height: 100vh;
-    padding: 10px;
-    background-color: #f3f3f4;
-    box-sizing: border-box;
-    align-items: center;
-    justify-content: center;
-}
-
-.title {
+.logo {
     text-align: center;
     margin: 10px;
 }
